Adds short break and long break modes
You can now switch to a short break or a long break after you finish a
pomodoro. This was done by adding a TIMER_MODE_CHANGE action that
changes the time_remaining_ms to the correct time for the given mode,
25m pomodoro, 5m short break, 10m long break.

Whenever you switch a mode, you automatically begin the timer for that
mode. The buttons to switch modes also reflect which one is active
whenever you click on that button.