Adds ability to sign out
You can now log out of the application. One wrinkled this functionality
posed was making sure that the csrf token updated appropriately.
Whenever you sign out, the session is destroyed and the csrf token is
destroyed too. This works fine with server rendering, since a new csrf
token is embeded on the next page load, but must be handled differently
in a SPA since the page does not refresh. To get around this, the new token
is now embedded in the sign out response, and the new token is
inserted into the meta tag.