Makes sessions fully functional
This commit does three things:

1) Whenever you load the page, the menu will correctly display either
Log Out, or Sign Up and Log In depending on whether you are logged in at
page load. This is done by checking for a data div created server side
that says whether or not you are logged in. This value is then added to
the redux store.

2) Whenever you log in, the session active state in redux is updated to
true, and the correct links are shown.

3) Whenever you log out, the session active state in redux is set to
false, and the correct links are shown.