Adds Timer Settings and allows them to be updated
This commit introduces the ability to change the timer settings for your
account. Pomodoro, short break, and long break lengths can all be
updated. To do this, the Timer Settings model was created, which is created
at the same time as a user. There is also now a
Api::TimerSettingsController which has show and update routes. The
SettingsForm and SettingsPage components show the settigns on the front
end.

If you log out, the timer goes back to default settings.