Updates timer settings in Redux after PUT request
Before, the app wasn't updating the timer settings in redux until I queried for them on
the main page. This change makes it so that they are updated with the
response from the PUT, and this will make it so you never see the wrong
timer settings in the Timer without needing a spinner.