Makes updating settings handle unauthenticated requests
This meant handling two cases:

1) User signed out in another tab, then tried to update settings
2) User's session expired

For 1, what happens is the CSRF token is refreshed in the other tab, but
not in the tab where the update is occuring. This triggers a CSRF 422
response. To make this distinguishable from other 422s due to AR
validation, a json response with invalid_authenticity_token: true is
rendered. This logic is contained in the new ApiController class. On the
client side, a method fetchAuthenticatedResource was added that does a
fetch with CSRF, and then catches any error and checks for that json
key. If it's present, it triggers a full page refresh, which is just an
easy way to handle this.

For 2, the controller will return a 401. In the
fetchAuthenticatedResource method, it also checks for 401s, and sets
session.active = false in the redux store. This causes the new
AuthenticatedComponentContainer, which surrounds routes that need to be
authenticated, to render a <Redirect> component instead of its children
(the other routes).