Adds and displays timer settings validations
On the backend, active record validations were added to the Timer
Settings to ensure they are present, and greater than 0. Validation
errors are displayed on the front end like before. The fetchWithCSRF
function has been modified to ensure that the body of the error messages
are parsed if they are present.