Adds ability to show saved tasks
Users can now see their tasks after a browser refresh. The index action
on the api/tasks_controller was implemented, and the corresponding
REQUEST_TASKS and RECEIVE_TASKS actions were implemented on the client
side.

This commit also makes the tasks related content only show if the user
is logged in.