Improves log out security, and caches tasks
After a log out or log in, the redux state is entirely reset to default.
This is done through a delegating root reducer that sets the state of the app
to undefined if it receives the SESSION_CHANGED action. This commit also
adds a test to make sure that tasks won't leak between users.

If there are tasks in the store, the app will now forgoe a request to
the server, and just use the local tasks.