Fixes csrf bug caused by switching to warden authenticate
Whenever I changed the sign in method to warden.authenticate(), I did
not know that this had an effect on the expected csrf token. It must
reset the token that the app expects. This was causing csrf invalid
errors after sign in, so I now return a new token in the response after
sign in.