Potentially fixes authenticated component problem
On mobile, whenever you log in with remember me checked, go to the
settings page (an authenticated page), then kill the app (killing the
session), you will be redirected to the login screen even though you are
still logged in. session.active will be set to false until you refresh,
providing a weird experience. This is caused because the default
session.active = false is set whenever mobile chrome revives the page.

To fix this, I added a wasVerified field to the session state. Before
you get redirected to login and treated as though you've been logged
out whenever you access an authenticated component, the app will check
to see if that has been remotely verified, and if not, will reach out to
the new users controller to check if you're signed in. If you are, the
page loads like normal, and if not, you get redirected to login.