Adds completing tasks towards pomodoro features
Whenever you click a task, go to its page, and complete a pomodoro, this
is recorded on the server as a pomodoro. These pomodoros are tracked in
redux and the completed count is incremented.

Instead of just keeping the count of completed pomodoros on the task
model, I put them on their own, pomodoro model, to make it easier in the
future to display summary statistics, like how many pomodoros you
completed today.