Adds completed count to tasks on task list
The app now pulls the pomodoros down whenever you go to the task page if
there are none in the cache, or if the only pomodoros were loaded for
the active task. The second case is necessary whenever you go home after
refreshing the page on a task page. To support this, this commit adds a
route for /api/pomodoros to the pomodoros controller.