Removes Timer from tasks list
Whenever a user is logged in, the home page no longer displays a timer.
Eventually this will be replaced with a timer in the footer, but for
now, it's just gone to make the site look a little better.

Commit also refactors Task model to add a method to mark the task as
complete, so that that logic is not located inside the
CreateTaskService.