Adds ability to mark daily goals accomplished
To do this on the front end, I used redux-json-api, to spare me from
writing any reducers. It ends up working out nicely, as redux-json-api
can manage the state in redux, and automatically update the goal record
in redux to be accomplished, showing the strikethrough on the front end.

To keep track of when a user's finished a daily goal, I added the
accomplishment model, which right now just has the created_at date, and
the goal foreign key. The Goal right now has an accomplished_today
method, which checks if there was an accomplishment during this UTC time
day, which isn't quite what we want. I will have to tailor that to the
user's real timezone in the future.