Refactors flexible service and update task service
Refactors flexible service to split Success and Failure results into
their own classes instead of using the @was_successful instance variable
and a bunch of conditionals. Refactors UpdateTaskService and Task model
to set completed_at time with #completed= instead of #complete. Makes
the service cleaner, no need for a conditional on what is contained in
the params passed to the service.