Adds ability to create excuses for daily goals
If you do not accomplish a daily goal, you can now list an excuse for
why you did not accomplish the goal.

To do this, there's now a paper icon that appears next to the checkbox
for each daily goal. Clicking on this goal launches a modal
  (ExcuseModal.js), which has a form (ExcuseForm.js), where you can add
  a description for the excuse. Eventually would like to make this an
  auto completed and editable list of different excuses per user, but
  that is not really necessary for my use case, at least now.

The excuse is posted to the server within the duck excuses.js. Right now
there is no error handling (of course). Hits the api/excuses controller.