Makes home page list existing excuses
Now shows icons on each daily goal with paper if there is an excuse for
a daily goal. Still will not allow you to edit an existing excuse, and
will allow for creation of more than one excuse per day, but this is a
decent start and will let users start to accumulate excuses to list
later.

Also adds a procfile with a release section so that I will not have to
run migrations manually when deploying the app.

Also deletes the persistent cookie csrf initializer. This did not appear
to be working properly with the version of Devise that I upgraded to. I
will maybe one day make authentication work better, but for now devises
defaults should be adaquete.