Adds ability to update excuses
You can now update the excuses that you've given to a daily goal. Adds
an update method to the ExcusesController, and adds an updateExcuse
method to the front end as well. The attributes for the excuse are retrieved
from the redux store and given to the ExcuseFrom as the initialValues
prop.