Adds ability to change volume
A much needed feature has been added in very raw form. Could use a
button that plays the sound so that people can see if the volume is what
they want. This has wide reaching changes, and hasn't been implemented
in a great way but I am a busy man at this point.